Moving InstanceFactory
